refactor: Add custom query & query suspense
