+++ src/pages/VerticalScreen.tsx
@@ -1,46 +1,13 @@
-import React from 'react';
-import {FlatList, SafeAreaView, StyleSheet} from 'react-native';
-import {useInfiniteQuery} from '@tanstack/react-query';
-import {photoAPI} from '../api/photoAPI';
-import {IPhoto} from '../types/IPhoto';
-import {VerticalItem} from '../components/VerticalItem';
-import {VerticalFooter} from '../components/VerticalFooter';
+import React, {Suspense} from 'react';
+import {SafeAreaView, Text} from 'react-native';
+import {VerticalList} from '../components/VerticalList';
 
 export function VerticalScreen() {
-  const {data, hasNextPage, fetchNextPage} = useInfiniteQuery<IPhoto[]>(
-    ['photos'],
-    ({pageParam}) => photoAPI.get(pageParam) as Promise<IPhoto[]>,
-    {
-      getNextPageParam: (lastPage, allPages) => {
-        return allPages.length + 1;
-      },
-    },
-  );
-
-  const loadMore = () => {
-    if (hasNextPage) {
-      fetchNextPage();
-    }
-  };
-
   return (
     <SafeAreaView>
-      <FlatList
-        style={styles.listWrap}
-        renderItem={VerticalItem}
-        ListFooterComponent={VerticalFooter}
-        onEndReached={loadMore}
-        onEndReachedThreshold={0.75}
-        data={data?.pages.flat()}
-      />
+      <Suspense fallback={<Text>loading...</Text>}>
+        <VerticalList />
+      </Suspense>
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  listWrap: {
-    paddingTop: 16,
-    paddingLeft: 16,
-    paddingRight: 16,
-  },
-});
